+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/DiagnosticTestGenerated.java
@@ -48,6 +48,18 @@ public class DiagnosticTestGenerated extends AbstractDiagnosticTest {
     public void testInternalHintsAreNotVisibleWithoutFriends() {
       runTest("compiler-tests/src/test/data/diagnostic/aggregation/InternalHintsAreNotVisibleWithoutFriends.kt");
     }
+
+    @Test
+    @TestMetadata("NonPublicContributionError.kt")
+    public void testNonPublicContributionError() {
+      runTest("compiler-tests/src/test/data/diagnostic/aggregation/NonPublicContributionError.kt");
+    }
+
+    @Test
+    @TestMetadata("NonPublicContributionWarning.kt")
+    public void testNonPublicContributionWarning() {
+      runTest("compiler-tests/src/test/data/diagnostic/aggregation/NonPublicContributionWarning.kt");
+    }
   }
 
   @Nested